requirements.txt
-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
async-generator==1.10; python_version >= '3.5'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backports.zoneinfo==0.2.1; python_version < '3.9'
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1; python_full_version >= '3.6.1'
charset-normalizer==2.0.11; python_version >= '3'
chromedriver-py==97.0.4692.71
click==8.0.3; python_version >= '3.6'
cryptography==36.0.1
distlib==0.3.4
django-environ==0.8.1
django-extensions==3.1.5
django-js-asset==1.2.2
django-mptt==0.13.4
django-treebeard==4.5.1
django==4.0.2
factory-boy==3.2.1; python_version >= '3.6'
faker==12.0.0; python_version >= '3.6'
filelock==3.4.2; python_version >= '3.7'
h11==0.13.0; python_version >= '3.6'
identify==2.4.7; python_version >= '3.7'
idna==3.3
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
outcome==1.1.0; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pathspec==0.9.0
pipdeptree==2.2.1
pipfile-sort==0.2.2
platformdirs==2.4.1; python_version >= '3.7'
plette==0.2.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==1.0.0; python_version >= '3.6'
pre-commit==2.17.0
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.7; python_version >= '3.6'
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-html==3.1.1; python_version >= '3.6'
pytest-metadata==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytest-selenium==2.0.1
pytest-sugar==0.9.4
pytest-variables==1.9.0
pytest==6.2.5
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0; python_version >= '3.6'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
selenium==4.1.0; python_version ~= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
sqlparse==0.4.2; python_version >= '3.5'
tenacity==6.3.1
termcolor==1.1.0
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.0; python_version >= '3.7'
tomlkit==0.9.0; python_version >= '3.6' and python_version < '4.0'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.19.0; python_version >= '3.6'
typing-extensions==4.0.1; python_version < '3.10'
urllib3[secure]==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
virtualenv==20.13.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
werkzeug==2.0.2
wsproto==1.0.0; python_full_version >= '3.6.1'
